refactor: ResultMapper를 클래스화한 Results 생성
